feat(cache checkpoint): cache checkpoint
